backend/requirements.txt
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
pandas==1.1.1
phonenumbers==8.12.9
python-dotenv
mysqlclient
